chore: move to new ESLint config format
ESLint is providing a new config format (typically called flat config).
The old config format will be deprecated in ESLint 9.0 (currently in
rc), and will be removed in ESLint 10. This commit migrate from the old
 config format at .eslintrc.js to the new flat config one at
 eslint.config.js.

In addition, 3 packages were added:

- @types/eslint: A dev dependency used to provide ESLint rules code
suggestion.
This dependency is used to provide linting through jsdoc for `styleRule`
 function. It has a very weak consumption (only 1 function relies on
 it), but it is a small types package. The old config file also relies
 on it (but does not install it as a npm package).

- globals: A dev dependency used to configure ESLint global linting
environment.
It is consumed by various other package, and already present on the
path. Although, some package depend on a very old version of globals
(babel), and with how terrible npm symlinking is, ESLint would import
version 11.x.x, with a tragic yet funny bug that will break ESLint.

- typescript-eslint: A dev dependency that provide first class
Typescript integration with ESLint.
It provides a convenient config function fully typed to create ESLint
flat config array. Additionally, provide the recommended
Typescript/ESLint settings that can be conveniently extended.

This commit also deleted 2 other dependencies:

- @typescript-eslint/eslint-plugin
- @typescript-eslint/parser

Both these two packages are no longer needed to be explicitly declared
as typescript-eslint already import them. And it will make better
decision on which version of them it should depend on.

chore: ESM support

package.json is changed to declare module (ESM), practically mark the
project as an ESM first class. Additionally, using ESM module to
configure Playwright and ESLint.
