+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "homepage": "https://fribbels.github.io/hsr-optimizer",
+  "type": "module",
   "dependencies": {
     "@ant-design/icons": "^5.2.6",
     "@js-sdsl/priority-queue": "^4.4.0",
@@ -63,19 +64,20 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/eslint": "^8.56.6",
     "@types/node": "^20.11.6",
     "@types/object-hash": "^3.0.6",
-    "@typescript-eslint/eslint-plugin": "^6.19.0",
-    "@typescript-eslint/parser": "^6.19.0",
     "@vitejs/plugin-legacy": "^5.3.0",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "eslint": "^8.56.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "gh-pages": "^6.1.0",
+    "globals": "^15.0.0",
     "jsdom": "^24.0.0",
     "terser": "^5.27.0",
     "typescript": "^5.3.3",
+    "typescript-eslint": "^7.4.0",
     "vite": "^5.2.6",
     "vite-tsconfig-paths": "^4.3.1",
     "vitest": "^1.4.0"
